Apply MovieNotFoundException to the MovieService layer.

+++ src/main/java/com/randyshreeves/videostreaming/exception/MovieNotFoundException.java
@@ -2,7 +2,7 @@ package com.randyshreeves.videostreaming.exception;
 
 public class MovieNotFoundException extends RuntimeException {
 
-    public MovieNotFoundException(int id) {
+    public MovieNotFoundException(Long id) {
         super("Movie with ID: " + id + " was not found.");
     }
 }
